Make genie-chrome amending testem instead of replacing it
Previous implementation was based on a full replacement on testem.js file in order to set configuration for Chrome in CI. This feature is amending the file instead, allowing to keep users modification while setting proper Chrome config. This will allow to have both `genie-chrome` and `genie-coverage` amending the file, for example. Distributing testem.js file in blueprint is not required anymore.